[IA64] bug fix in privop.c
Bug fix: an ia64 register is 64bits!

+++ xen/arch/ia64/xen/privop.c
@@ -409,7 +409,7 @@ IA64FAULT priv_mov_from_rr(VCPU *vcpu, INST64 inst)
 {
 	UINT64 val;
 	IA64FAULT fault;
-	int reg;
+	UINT64 reg;
 	
 	reg = vcpu_get_gr(vcpu,inst.M43.r3);
 	if (privify_en && inst.M43.r1 > 63) {
